Move SystemExitHelper to lib package
