+++ org-code-javabuilder/lib/src/main/java/dev/javabuilder/LocalSystemExitHelper.java
@@ -1,6 +1,6 @@
 package dev.javabuilder;
 
-import org.code.protocol.SystemExitHelper;
+import org.code.javabuilder.SystemExitHelper;
 
 public class LocalSystemExitHelper implements SystemExitHelper {
